refactor: move addItemToInventory to its own file
